Add bundled favicon logic & fix nits
